requirements.txt
torch>=2.6.0
torchvision>=0.21.0
albumentations>=1.4.0
opencv-python>=4.10.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.13.0
grad-cam>=1.5.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.66.0
efficientnet-pytorch>=0.7.1
timm>=1.0.0